Course_recommendation_system/requirements.txt
Flask==3.1.0
numpy==2.2.0
pandas==2.2.3
scikit_learn==1.4.2
scikit_surprise==1.1.4
